% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDRMapp.R
\name{clustDRMapp}
\alias{clustDRMapp}
\title{launch the shiny app for an easier use of the package}
\usage{
clustDRMapp()
}
\description{
launch the shiny app for an easier use of the package
}
\details{
the shiny app made for an easy use of the functionalities 
of the clustDRM package. It can be launched using command clustDRMapp(). 
It imports the data (in csv format),
performs the clustering on it (for monotone patterns using E2
test, and for general patterns using ORICC2 and MCT). It also
can estimate EDp for different dose-response curves using 
appropriate models. Plotting dose-response curves are also
possible for any of these operations. A simulation tab will help
to decide about the design of the study using a simulation study.
}
\author{
Vahid Nassiri and Yimer Wasihun.
}
